processing audio from TCDTIMIT

+++ utils/copyFilesOfType.py
@@ -1,6 +1,9 @@
-#!/usr/bin/python3
+from __future__ import print_function
 import os, re, os.path
-import os, shutil, sys
+import shutil, sys
+import argparse
+import filecmp
+from tqdm import tqdm
 
 protocolPattern = re.compile('r&apos^\w+://&apos')
 
@@ -71,7 +74,7 @@ def query_yes_no(question, default="yes"):
 
     while True:
         sys.stdout.write(question + prompt)
-        choice = input().lower()
+        choice = raw_input().lower()
         if default is not None and choice == '':
             return valid[default]
         elif choice in valid:
@@ -80,9 +83,8 @@ def query_yes_no(question, default="yes"):
             sys.stdout.write("Please respond with 'yes' or 'no' "
                              "(or 'y' or 'n').\n")
 
-
-def copyFilesOfType(srcDir, dstDir, extension):
-    print("Source Dir: %s, Destination Dir: %s, Extension: %s" %(srcDir, dstDir, extension))
+def copyFilesOfType(srcDir, destDir, extension, interactive=True):
+    print("Source Dir: %s, Destination Dir: %s, Extension: %s" %(srcDir, destDir, extension))
 
     src = []
     dest = []
@@ -90,26 +92,41 @@ def copyFilesOfType(srcDir, dstDir, extension):
         for file_ in files:
             if file_.endswith(extension):
                 srcPath = os.path.join(root, file_)
-                relSrcPath = relpath(srcDir, srcPath)
-                destPath = os.path.join(dstDir, relSrcPath)
-                print("copying from : %s to \t\t %s" % (srcPath, destPath))
+                relSrcPath = relpath(srcDir, srcPath).lstrip("../")
+                # print(relSrcPath)
+                destPath = os.path.join(destDir, relSrcPath)
+                # print("copying from : %s to \t\t %s" % (srcPath, destPath))
                 src.append(srcPath)
                 dest.append(destPath)
 
-    print(len(src), " files will be copied")
-    if query_yes_no("Are you sure you want to peform these operations?","yes"):
-        for i in range(len(src)):
+    print("Example: copying ", src[0], "to:", dest[0])
+    print(len(src), " files will be copied in total")
+
+    if ( interactive and (not query_yes_no("Are you sure you want to peform these operations?","yes"))):
+        print ("Not doing a thing.")
+    else:
+        for i in tqdm(range(len(src))):
             if (not os.path.exists(os.path.dirname(dest[i]))):
                 os.makedirs(os.path.dirname(dest[i]))
+
+            # if file already exists, skip
+            if ( os.path.exists(dest[i]) ):
+                # if (filecmp.cmp(src[i],dest[i]) == True):  #this reads all the files anyway, so doesn't make sense
+                    # print("file exists, skipping")
+                    # print("src: ", src[i],' Dst: ', dest[i])
+                continue
             shutil.copy(src[i], dest[i])
-    print("Done.")
+
+        print("Done.")
+
+    return 0
 
 
 if __name__ == '__main__':
-    srcDir = sys.argv[1]
-    dstDir = sys.argv[2]
-    extension = sys.argv[3]
-    copyFilesOfType(srcDir,dstDir,extension)
+    if __name__ == '__main__':
+        srcDir = sys.argv[1]
+        dstDir = sys.argv[2]
+        type = sys.argv[3]
+        copyFilesOfType(srcDir,dstDir,type, interactive=True)
 
-
-# example usage: python3 fileDirOps.py ~/Documents/Dropbox/_MyDocs/_ku_leuven/Master_2/Thesis/convNets/code /media/matthijs/TOSHIBA_EXT/TCDTIMIT/zzzNPZmodels ".npz"
+# example usage: python3 copyFilesOfType.py ~/Documents/Dropbox/_MyDocs/_ku_leuven/Master_2/Thesis/convNets/code /media/matthijs/TOSHIBA_EXT/TCDTIMIT/zzzNPZmodels ".npz"
